Moved from github dependency to npm
