+++ android/app/build.gradle
@@ -222,7 +222,7 @@ dependencies {
         implementation jscFlavor
     }
 
-    implementation files("../../node_modules/react-native-lightning/android/libs/Lndmobile.aar")
+    implementation files("../../node_modules/@synonymdev/react-native-lightning/android/libs/Lndmobile.aar")
 }
 
 // Run this once to be able to run the application with BUCK
