+++ package.json
@@ -52,7 +52,7 @@
     "react-native-gesture-handler": "^1.10.3",
     "react-native-haptic-feedback": "^1.11.0",
     "react-native-keychain": "^7.0.0",
-    "react-native-lightning": "git+ssh://git@github.com/synonymdev/react-native-lightning",
+    "@synonymdev/react-native-lightning": "0.0.29",
     "react-native-modal": "^11.10.0",
     "react-native-qrcode-svg": "^6.1.1",
     "react-native-randombytes": "^3.6.1",
